preview componet fixes
